Feature/sharing

+++ src/store/driveStore.ts
@@ -68,33 +68,54 @@ export const signIn = (id: string) => {
   gapi.load("auth2", () => {
     gapi.auth2.getAuthInstance().currentUser.listen((val) => {
       if (val.getId() === null) {
-        /* TODO: there is no user logged in, try and retrieve it. It might be public. */
         state.loggedIn = false;
-        const exampleDataset = examples.find(e => e.title.toLowerCase().replaceAll(" ", "-") === router.currentRoute.params.fileId);
-        if (router.currentRoute.name === "Explore" && !exampleDataset) {
-          router.push({ name: "Home" });
-        }
+        directLinkDownloadData();
       }
     });
   });
 };
 
 const directLinkDownloadData = () => {
-  if (router.currentRoute.params.fileId && exploreState.uploadedFile === null) {
-    const file = state.files.find((_) => _.id === router.currentRoute.params.fileId);
-    if (file) {
-      const configFile = state.files.find(
-        (_) => _.name === `${file.name}.json`
+  const fileId = router.currentRoute.params.fileId;
+  if (fileId) {
+    const configFile = state.files.find((_) => _.id === fileId);
+    if (configFile) {
+      const file = state.files.find(
+        (_) => _.name === `${configFile.name!.split(".").slice(0, -1).join(".")}`
       )!;
       const geojsonFile = state.files.find(
         (_) => _.name === `${file.name}.geojson.json`
       )!;
-      downloadUserUpload({ file, configFile, geojsonFile });
+      downloadUserUpload({ file, configFile, geojsonFile }, true);
     } else {
-      /* TODO: Try to download this file even though the user did not create it. */
-      const exampleDataset = examples.find(e => e.title.toLowerCase().replaceAll(" ", "-") === router.currentRoute.params.fileId);
+      const exampleDataset = examples.find(e => e.title.toLowerCase().replaceAll(" ", "-") === fileId);
       if (!exampleDataset) {
-        router.push({ name: "404" });
+        gapi.load("client", () => {
+          gapi.client.setApiKey(process.env.VUE_APP_LOGGED_OUT_USER_API_KEY);
+          gapi.client.load("drive", "v3", async () => {
+            downloadFile(fileId).then(resp => {
+              const body = JSON.parse(resp);
+              gapi.client.drive.files.get({
+                fileId: body.ids.file
+              }).execute((resp) => {
+                downloadUserUpload({
+                  file: {
+                    id: body.ids.file,
+                    name: resp.result.name
+                  },
+                  configFile: {
+                    id: fileId
+                  },
+                  geojsonFile: {
+                    id: body.ids.geojsonFile
+                  }
+                }, false);
+              });
+            }).catch(() => {
+              router.push({ name: "404" });
+            });
+          });
+        });
       }
     }
   }
@@ -147,6 +168,31 @@ export const refreshFiles = (callback?: () => void | undefined) => {
   });
 };
 
+export const updateShared = (files: gapi.client.drive.File[], shared: boolean) => {
+  files.forEach(file => {
+    if (shared) {
+      return gapi.client.drive.permissions.create({
+        resource: {
+          type: "anyone",
+          role: "reader"
+        },
+        fileId: file.id!
+      }).execute(() => null);
+    } else {
+      return gapi.client.drive.permissions.delete({
+        permissionId: "anyoneWithLink",
+        fileId: file.id!
+      }).execute(() => null);
+    }
+  });
+  files.forEach(f => {
+    const file = state.files.find(_ => _.id === f.id);
+    if (file) {
+      file.shared = shared;
+    }
+  });
+}
+
 export const downloadFile = (fileId: string) => {
   return gapi.client.drive.files
     .get({ fileId, alt: "media" })
@@ -155,12 +201,14 @@ export const downloadFile = (fileId: string) => {
     });
 }
 
-export const uploadFile = (data: string, mimeType: string, name: string, callback?: (fileName: string) => void | undefined) => {
-  const metadata = {
+export const uploadFile = (data: string, mimeType: string, name: string, id?: string | undefined, callback?: (file: gapi.client.drive.File) => void | undefined) => {
+  const metadata: { name: string, mimeType: string, parents?: string[] | undefined } = {
     name,
-    mimeType,
-    parents: [state.folderId]
+    mimeType
   };
+  if (!id) {
+    metadata.parents = [state.folderId!];
+  }
   const boundary = "TableAndMap";
   const delimiter = "\r\n--" + boundary + "\r\n";
   const closeDelim = "\r\n--" + boundary + "--";
@@ -173,26 +221,28 @@ export const uploadFile = (data: string, mimeType: string, name: string, callbac
     data + "\r\n" +
     closeDelim;
   gapi.client.request({
-    path: "https://www.googleapis.com/upload/drive/v3/files",
-    method: "POST",
+    path: "https://www.googleapis.com/upload/drive/v3/files" + (id ? `/${id}` : ""),
+    method: id ? "PATCH" : "POST",
     params: { uploadType: "multipart", fields: "*" },
     headers: {
       "Content-Type": "multipart/related; boundary=" + boundary
     },
     body: multipartRequestBody
   }).execute((resp: any) => {
-    if (callback) {
-      callback(resp.id);
-    }
-    const existingFiles = state.files.filter(_ => _.name === name);
-    if (existingFiles.length) {
-      existingFiles.forEach(existingFile => {
-        gapi.client.drive.files.delete({
-          fileId: existingFile.id!
-        }).execute(() => null);
-      });
+    if (!id) {
+      const existingFiles = state.files.filter(_ => _.name === name);
+      if (existingFiles.length) {
+        existingFiles.forEach(existingFile => {
+          gapi.client.drive.files.delete({
+            fileId: existingFile.id!
+          }).execute(() => null);
+        });
+      }
     }
     state.files = state.files.filter(_ => _.name !== name).concat(resp);
+    if (callback) {
+      callback(resp);
+    }
   });
 }
